// apps/api/src/shared/application/auth/permissions.d.ts
export type Permission = "POST_CREATE" | "POST_UPDATE" | "POST_PUBLISH" | "POST_ARCHIVE" | "POST_DRAFT" | "USER_CREATE" | "USER_UPDATE" | "USER_CHANGE_ROLE" | "USER_CHANGE_PASSWORD" | "USER_CHANGE_STATUS" | "AUTHOR_CREATE" | "AUTHOR_UPDATE" | "AUTHOR_CHANGE_STATUS" | "CATEGORY_CREATE" | "CATEGORY_UPDATE";
//# sourceMappingURL=permissions.d.ts.map